refactor(자동차 이름 수, 시도 횟수): 제한값 10이하로 변경
