+++ src/main/java/racingcar/game/io/InputTryTimes.java
@@ -1,7 +1,6 @@
 package racingcar.game.io;
 
 import java.util.Scanner;
-import java.util.regex.Pattern;
 import racingcar.game.io.error.TryTimeInputErrorHandler;
 import racingcar.game.io.print.InputPrint;
 
@@ -17,38 +16,7 @@ public class InputTryTimes {
         do {
             InputPrint.printMessageForInputTryTimes();
             input = scanner.nextLine();
-        } while (!isValidInput(input));
+        } while (!errorHandler.isValidInput(input));
         return Integer.parseInt(input);
-    }
-
-    private boolean isValidInput(String input) {
-        if (!isNumber(input)) {
-            errorHandler.printNotNumberErrorMessage();
-            return false;
-        }
-        if (!isCorrectRange(input)) {
-            errorHandler.printInvalidRangeNumberErrorMessage();
-            return false;
-        }
-        return true;
-    }
-
-    private boolean isCorrectRange(String input) {
-        if (input.isEmpty()) {
-            return false;
-        }
-        if (5 < input.length()) {
-            return false;
-        }
-        int tryTimes = Integer.parseInt(input);
-        if (!(1 <= tryTimes && tryTimes <= 1000)) {
-            return false;
-        }
-        return true;
-    }
-
-    private boolean isNumber(String input) {
-        String pattern = "^[0-9]+$";
-        return Pattern.matches(pattern, input);
     }
 }
