+++ src/main/java/racingcar/game/io/error/print/CarNameInputErrorPrint.java
@@ -1,4 +1,4 @@
-package racingcar.game.io.error;
+package racingcar.game.io.error.print;
 
 public class CarNameInputErrorPrint {
     public static final String MORE_THAN_ONE_CAR_NAME_ERROR_MESSAGE
@@ -10,7 +10,7 @@ public class CarNameInputErrorPrint {
     public static final String CAR_NAME_LENGTH_ERROR
         = "[ERROR] 자동차 이름은 1자 이상 5자 이하만 가능합니다.";
     public static final String NUMBER_OF_CAR_MORE_THAN_MAX_ERROR_MESSAGE
-        = "[ERROR] 경주할 자동차 수는 100대 이하여야 합니다.";
+        = "[ERROR] 경주할 자동차 수는 10대 이하여야 합니다.";
     private static final String DUPLICATE_NAME_ERROR_MESSAGE
         = "[ERROR] 중복된 이름이 있습니다.";
